add some exception filter and modify update logic

+++ mongo/homework.py
@@ -1,42 +1,43 @@
 from . import engine
 from .course import Course
+from datetime import datetime
 
 __all__ = ['HomeWork']
 
 
 class HomeWork:
     @staticmethod
-    def add_hw(course_name,markdown , hw_name, start, end, problem_ids, scoreboard_status):
-       #query db check the hw doesn't exist       
+    def add_hw(course_name, markdown, hw_name, start, end, problem_ids,
+               scoreboard_status):
+        #query db check the hw doesn't exist
         course = engine.Course.objects(course_name=course_name).first()
         course_id = course.id
         students = course.students
-        homework = engine.Homework.objects.get(course_id=str(course_id),
-                                               name=hw_name)
-        if (homework is not none):
+        homework = engine.Homework.objects(course_id=str(course_id),
+                                           name=hw_name)
+        if (len(homework) != 0):
             raise FileExistsError
 
         homework = engine.Homework(name=hw_name,
                                    course_id=str(course_id),
                                    problem_ids=problem_ids)
-        homework.duration.start = start
-        homework.duration.end = end
-        homework.markdown = markdown
-        homework.scoreboard_status = scoreboard_status
+        homework.duration.start = datetime.now() if start is None else start
+        homework.duration.end = datetime.now() if end is None else end
+        homework.markdown = '' if markdown is None else markdown
+        homework.scoreboard_status = 0 if scoreboard_status is None else scoreboard_status
         #init student status
-        userIds = {}
-        userProblems = {}
-        for problem_id in problem_ids:
-            userProblems[str(problem_id)] = {
-                "score": 0,
-                "problemStatus": 1,
-                "submissonIds": []
-            }
-
+        user_ids = {}
+        user_problems = {}
+        if (problem_ids is not None):
+            for problem_id in problem_ids:
+                user_problems[str(problem_id)] = {
+                    "score": 0,
+                    "problemStatus": 1,
+                    "submissonIds": []
+                }
         for key in students:
-            userIds[key] = userProblems
-
-        homework.student_status = userIds
+            user_ids[key] = user_problems
+        homework.student_status = user_ids
         homework.save()
         #get homeworkId then store in the correspond course
         homework = engine.Homework.objects(
@@ -50,46 +51,54 @@ class HomeWork:
         return homework
 
     @staticmethod
-    def update(course_name, markdown, hw_name, new_hw_name, start, end, problem_ids, scoreboard_status):
+    def update(course_name, markdown, hw_name, new_hw_name, start, end,
+               problem_ids, scoreboard_status):
         #check course exist
         course = engine.Course.objects(course_name=course_name).first()
-        if (course is none):
+        if (course is None):
             raise FileNotFoundError
         course_id = course.id
         students = course.students
-        #check the new_name hasn't been use in this course
-        homework = engine.Homework.objects.get(course_id=str(course_id),
-                                               name=new_hw_name)
-        if (homework is not none):
-            raise FileExistsError
-
-        #get the homework and update it's field
+        #get the homework
         homework = engine.Homework.objects.get(course_id=str(course_id),
                                                name=hw_name)
-        if (start != 0):
-            homework.duration.start = start
-        if (end != 0):
-            homework.duration.end = end
-        homework.scoreboard_status = scoreboard_status
-        #if problemid exist then delete ,else add it in list
-        userProblems = {}
-        userIds = {}
-        #傳進來的problem_ids應該只有要新增/刪除的
-        for x in problem_ids:
-            if (x in homework.problem_ids):
-                homework.problem_ids.remove(x)
-                for userId in homework.student_status:
-                    homework.student_status[userId].pop(x)
+        #check the new_name hasn't been use in this course
+        if new_hw_name is not None:
+            result = engine.Homework.objects(course_id=str(course_id),
+                                             name=new_hw_name)
+            if (len(result) != 0):
+                raise FileExistsError
             else:
-                homework.problem_ids.append(x)
-                for key in students:
-                    homework.student_status[key][x] = {
-                        "score": 0,
-                        "problemStatus": 1,
-                        "submissonIds": []
-                    }
-        homework.markdown = markdown
-        homework.name = new_hw_name
+                homework.name = new_hw_name
+
+        #update fields
+        if (start is not None):
+            homework.duration.start = start
+        if (end is not None):
+            homework.duration.end = end
+        if (scoreboard_status is not None):
+            homework.scoreboard_status = scoreboard_status
+        #if problemid exist then delete ,else add it in list
+        user_problems = {}
+        user_ids = {}
+        #傳進來的problem_ids應該只有要新增/刪除的
+        if (problem_ids is not None):
+            for id in problem_ids:
+                if (id in homework.problem_ids):
+                    homework.problem_ids.remove(id)
+                    for userId in homework.student_status:
+                        homework.student_status[userId].pop(id)
+                else:
+                    homework.problem_ids.append(id)
+                    for key in students:
+                        homework.student_status[key][id] = {
+                            "score": 0,
+                            "problemStatus": 1,
+                            "submissonIds": []
+                        }
+
+        if markdown is not None:
+            homework.markdown = markdown
         homework.save()
         return homework
 
@@ -100,8 +109,10 @@ class HomeWork:
         course_id = course.id
         homework = engine.Homework.objects(
             course_id=str(course_id),
-            homewor_name=hw_name,
-        ).first()        
+            name=hw_name,
+        ).first()
+        if (homework is None):
+            raise FileNotFoundError
         homework.delete()
         course.save()
         return homework
